feat: migrate authentication logic to new auth app and restructure project settings
