+++ backend/auth/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AuthConfig(AppConfig):
+    name = "auth"
+    label = "api_auth"
