+++ backend/auth/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from auth.views import LoginView, RegisterView
+
+urlpatterns = [
+    path("register/", RegisterView.as_view()),
+    path("login/", LoginView.as_view()),
+]
