+++ backend/config/urls.py
@@ -27,5 +27,6 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/token/", TokenObtainPairView.as_view()),
     path("api/token/refresh/", TokenRefreshView.as_view()),
+    path("api/", include("auth.urls")),
     path("api/", include("users.urls")),
 ]
